Random Number by Seed
To provide multiplayer playablitiy, I want to create a seed for games. This will be a number that the game uses to generate a sequence of numbers. By providing the same seed, the sequence will be 100% identical, allowing players to play same game on different clients. Bingo cards will still be random regardless of the seed. but the order in which the numbers are drawn will remain consistent.

I have also created temporary test, which will later be replaced by the main game.